requirements.txt
paho-mqtt==1.3.1
requests==2.4.3
urllib3==1.9.1
